Rename DataSeries to AnimationCurve, move to separate file
